refactor(app): split the control-plane contract, evidence and gateway roots
Move the per-capability prepare handlers and the invocations they build to
control/contract/reads.rs, the retained-bundle store with its grant-rechecking
read to control/evidence/store.rs, the screenshot encoding to
control/evidence/image.rs, the in-process action and read door to
control/gateway/local_action.rs, and the control-trace walk with
service_replay_trace to control/gateway/trace_replay.rs.

Every body moves unchanged: the statement-line multiset of each root at the
campaign base equals that of the root plus its siblings. The only widenings
are pub(super) marks, and each root re-exports EventsReadInvocation,
RawScreenshot, ScreenshotPixels and EvidenceStore at the visibility they had,
so no caller's path changes. No check in ObserverContract::prepare or in
either dispatcher moves.

Production lines: contract.rs 1,718 -> 1,261, evidence.rs 1,708 -> 1,221,
gateway.rs 1,686 -> 1,219; the five siblings are 514, 327, 219, 282 and 234.
The three entries leave size-baseline.txt and !budget falls 45,739 -> 40,383.

Campaign #367 child S5; issue #372.
