+++ crates/guards/size-baseline.txt
@@ -257,7 +257,11 @@
 # because a sub-threshold file with a recorded ceiling is a permission nobody
 # spends, and the budget follows them down by the same 1,297. Nothing was
 # deleted; the multiset proof is in PR #399's body.
-!budget 9264
+# Raised for the same branch: an `Mt5Event::OpeningPage` variant and one
+# branch in the block state machine deciding whether a block settles a
+# request. The session loop is the only place that knows what a
+# `history_start` opened; the slicing itself lives in the bridge.
+!budget 4142
 
 crates/app/src/app.rs 769
 
@@ -369,8 +373,5 @@ crates/app/src/app.rs 769
 # Pure moves, with `the_journal_bytes_are_fixed` and
 # `the_report_numbers_are_fixed` unchanged on both sides of it.
 
-crates/app/src/control/gateway.rs 1696
 crates/orderflow/src/projection.rs 1850
-crates/app/src/control/contract.rs 1718
-crates/app/src/control/evidence.rs 1708
 crates/orderflow/src/config.rs 1523
